adk_simple/requirements.txt
google-adk @ git+https://github.com/google/adk-python.git@main
google-cloud-aiplatform[adk,agent-engines]>=1.126.1
google-cloud-storage>=3.0.0
pillow>=12.0.0
python-dotenv>=1.2.1
requests>=2.32.5